Fully fix keyboard backlight
Keyboard backlight needs to be on when, and only when, both the screen
is on and the slider is open.

 * Create a thread to listen for slider events and notify the light
   service when the state changes.

 * The light service tracks screen and slider state and sets the
   keyboard backlight appropriately when either changes.

Note services do not seem to have any exit path, so there is no need
to track the slider thread or provide any mechanism for stopping it.
The only consideration is that the callback function (onSliderChanged)
is called from the slider thread, so it must be protected with a lock.

Change-Id: If73df9c3ff85496fbdc01b6dbff38afec60a728d